requirements.txt
Flask==3.0.2
Requests==2.31.0
langchain==0.1.6
langchain_community==0.0.20
langchain_google_genai==0.0.8
logzero==1.7.0
pandas==1.5.3
protobuf==4.23.4
python-dotenv==1.0.1
pydantic==1.10.9
docarray==0.40.0
faiss-cpu
gunicorn
psycopg2